S/CG: Fix loss of a fractional attentuation part

+++ source/client/snd_public.h
@@ -99,7 +99,7 @@ public:
 
 	virtual void StartLocalSound( const char *name, float fvol ) = 0;
 	virtual void StartLocalSound( sfx_s *sfx, float fvol ) = 0;
-	virtual void AddLoopSound( sfx_s *sfx, int entNum, float fvol, int attenuation ) = 0;
+	virtual void AddLoopSound( sfx_s *sfx, int entNum, float fvol, float attenuation ) = 0;
 
 	virtual void RawSamples( unsigned samples, unsigned rate, uint16_t width, uint16_t channels, const uint8_t *data, bool music ) = 0;
 	virtual void PositionedRawSamples( int entNum, float fvol, float attenuation, unsigned samples,
